Break the compiler out into component ruby classes
