+++ lib/bootstrap_email.rb
@@ -18,6 +18,7 @@ require_relative 'bootstrap-email/adapters/string_adapter'
 require_relative 'bootstrap-email/compiler'
 require_relative 'bootstrap-email/sass_cache'
 require_relative 'bootstrap-email/version'
+Dir[File.join(__dir__, 'bootstrap-email/components', '*.rb')].each { |file| require_relative file }
 
 if defined?(Rails)
   require_relative 'bootstrap-email/rails/premailer_railtie'
